changed the way we access attributes and abilities
why:
before these changes we used the localized label in order to set some data attributes in the dom and access the proper object in the actor.data property.
In order to implement a permanent combination of attribute and ability, we a needed a key that would not break when users change their language in the settings.

what:
replaced all localize methods where we actually just need the key. the UI is still localized correctly
* also added some more UI localization in general (chat message and difficulty dialog)